+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * Fix jack-in from inside of remote buffers.
 * [#2454](https://github.com/clojure-emacs/cider/pull/2454): Fix erratic inspector behavior when multiple REPLs are connected
+* [#2467](https://github.com/clojure-emacs/cider/pull/2467): Make generic cider ops use any available nrepl connection
 
 ## 0.18.0 (2018-09-02)